Added classes endpoint. Using custom course id instead of course code.
Course Code is same for eth, epj and ela variants of a course. But these
need to be treated as separate things. Therefore Course ID is now used
instead which is a combination of course code and course type.

Classes endpoint returns all the classes taught by a particular faculty
for a particular course.