Minor changes.
Displaying selected faculties and not selected faculties differently.
When dnd is implemented this will make it more obvious that the selected
faculty rows can be reordered.

In Timetables section now rendering schedules above the filled out
timetable. This is because when the table is filled out it becomes bigger
and if it was above the schedules tables, then the schedules tables would
be pushed down. Instead this way the schedules tables stays at the same
height.