Wrote basic functions that will enable faster timetable generation.
While populateSlotCombination needs to be tested more, so far it looks
fine and generates all possible schedules given a slot combination to
follow.

getSlotCombinations doesn't get all slot combinations but works gets
enough of them that it's okay for now.

Priority should be switching the whole application to use these functions
instead of the old ones.