Figured out html2canvas. Semi-Major refactoring of Timetables.
Class Table is now a seperate component so as to allow us to easily
render each class table and generate a pdf that user can download.

Tried using react-pdf for making pdf but react-pdf doesn't render
any regular react dom elements. For now only option is to use
html2canvas on the timetable and each Class Table, and then maybe
use jsPDF to make a pdf. This would mean that none of the text is
selectable so some alternative should also be figured out asap.